// src/services/enums/races.ts
export enum ERaces {
    Nothing = 'Не выбрано',
    Orsimer = 'Орк',
    Nord = 'Норд',
    Redguard = 'Редгард',
    Altmer = 'Высокий эльф',
    Dunmer = 'Темный эльф',
    Breton = 'Бретон',
    Bosmer = 'Лесной эльф',
    Khajiit = 'Каджит',
    Imperial = 'Имперец',
    Argonian = 'Аргонианин'
}